Permissions prevents deletion. Permissions done for now.

+++ backend/src/resolvers/Mutation.js
@@ -48,8 +48,18 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // find the item
-    const item = await ctx.db.query.item({ where }, info, `{ id, title }`);
+    const item = await ctx.db.query.item(
+      { where },
+      `{ id, title, user { id }}`
+    );
     // check ownership of item/permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permissions to delete this item!");
+    }
     // delete it
     return ctx.db.mutation.deleteItem({ where }, info);
   },
